+++ spec/spec_helper.rb
@@ -7,6 +7,5 @@ Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
 require_relative "../lib/hotel"
 require_relative "../lib/room"
-require_relative "../lib/calendar"
 require_relative "../lib/reservation"
 # Require_relative your lib files here!
